requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cycler==0.10.0
Django==3.1.2
django-filter==2.4.0
djangorestframework==3.12.1
kiwisolver==1.2.0
Markdown==3.3.3
matplotlib==3.3.2
numpy==1.19.2
pandas==1.1.3
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
scipy==1.5.3
seaborn==0.11.0
six==1.15.0
sqlparse==0.4.1
